# Make hyriseBenchmarkX loading faster