--- a/scripts/compare_benchmarks.py
+++ b/scripts/compare_benchmarks.py
@@ -337,13 +337,15 @@ if github_format:
     green_control_sequence = colored("", "green")[0:5]
     red_control_sequence = colored("", "red")[0:5]
 
-    table_string_reformatted = "<details>\n" + \
-                               "<summary>Configuration Overview - click to expand</summary>\n\n" + \
-                               "```diff\n" + \
-                               create_context_overview(old_data, new_data, github_format) + \
-                               "```\n" + \
-                               "</details>\n\n" + \
-                               "```diff\n"
+    table_string_reformatted = (
+        "<details>\n"
+        + "<summary>Configuration Overview - click to expand</summary>\n\n"
+        + "```diff\n"
+        + create_context_overview(old_data, new_data, github_format)
+        + "```\n"
+        + "</details>\n\n"
+        + "```diff\n"
+    )
     for line in table_string.splitlines():
         if (green_control_sequence + "+" in line) or ("| Sum " in line and green_control_sequence in line):
             table_string_reformatted += "+"