--- a/src/benchmarklib/abstract_table_generator.cpp
+++ b/src/benchmarklib/abstract_table_generator.cpp
@@ -39,101 +39,112 @@ void AbstractTableGenerator::generate_and_store() {
   /**
    * Sort tables if a sort order was defined by the benchmark
    */
-  const auto& sort_order_by_table = _sort_order_by_table();
-  if (!sort_order_by_table.empty()) {
-    std::cout << "- Sorting tables" << std::endl;
+  {
+    const auto& sort_order_by_table = _sort_order_by_table();
+    if (!sort_order_by_table.empty()) {
+      std::cout << "- Sorting tables" << std::endl;
 
-    for (const auto& [table_name, column_name] : sort_order_by_table) {
-      auto& table = table_info_by_name[table_name].table;
-      const auto sort_mode = SortMode::Ascending;  // currently fixed to ascending
-      const auto sort_column_id = table->column_id_by_name(column_name);
-      const auto chunk_count = table->chunk_count();
+      // We do not use JobTasks here (and in the rest of this file) because we want this part to be multi-threaded even if
+      // Hyrise uses no scheduler.
+      auto threads = std::vector<std::thread>{};
+      for (const auto& [table_name, column_name] : sort_order_by_table) {
+        threads.emplace_back([&] {
+          auto& table = table_info_by_name[table_name].table;
+          const auto sort_mode = SortMode::Ascending;  // currently fixed to ascending
+          const auto sort_column_id = table->column_id_by_name(column_name);
+          const auto chunk_count = table->chunk_count();
 
-      // Check if table is already sorted
-      auto is_sorted = true;
-      resolve_data_type(table->column_data_type(sort_column_id), [&](auto type) {
-        using ColumnDataType = typename decltype(type)::type;
+          // Check if table is already sorted
+          auto is_sorted = true;
+          resolve_data_type(table->column_data_type(sort_column_id), [&](auto type) {
+            using ColumnDataType = typename decltype(type)::type;
 
-        auto last_value = std::optional<ColumnDataType>{};
-        for (auto chunk_id = ChunkID{0}; chunk_id < chunk_count; ++chunk_id) {
-          const auto& segment = table->get_chunk(chunk_id)->get_segment(sort_column_id);
-          segment_with_iterators<ColumnDataType>(*segment, [&](auto it, const auto end) {
-            while (it != end) {
-              if (it->is_null()) {
-                if (last_value) {
-                  // NULLs should come before all values
-                  is_sorted = false;
-                  break;
+            auto last_value = std::optional<ColumnDataType>{};
+            for (auto chunk_id = ChunkID{0}; chunk_id < chunk_count; ++chunk_id) {
+              const auto& segment = table->get_chunk(chunk_id)->get_segment(sort_column_id);
+              segment_with_iterators<ColumnDataType>(*segment, [&](auto it, const auto end) {
+                while (it != end) {
+                  if (it->is_null()) {
+                    if (last_value) {
+                      // NULLs should come before all values
+                      is_sorted = false;
+                      break;
+                    }
+
+                    ++it;
+                    continue;
+                  }
+
+                  if (!last_value || it->value() >= *last_value) {
+                    last_value = it->value();
+                  } else {
+                    is_sorted = false;
+                    break;
+                  }
+
+                  ++it;
                 }
-
-                ++it;
-                continue;
-              }
-
-              if (!last_value || it->value() >= *last_value) {
-                last_value = it->value();
-              } else {
-                is_sorted = false;
-                break;
-              }
-
-              ++it;
+              });
             }
           });
-        }
-      });
 
-      if (is_sorted) {
-        std::cout << "-  Table '" << table_name << "' is already sorted by '" << column_name << "' " << std::endl;
-        const SortColumnDefinition sort_column{sort_column_id, sort_mode};
+          if (is_sorted) {
+            auto output = std::string{"-  Table '"} + table_name + "' is already sorted by '" + column_name + "'\n";
+            std::cout << output << std::flush;
+            const SortColumnDefinition sort_column{sort_column_id, sort_mode};
 
-        if (_all_chunks_sorted_by(table, sort_column)) continue;
+            if (_all_chunks_sorted_by(table, sort_column)) return;
 
-        for (auto chunk_id = ChunkID{0}; chunk_id < chunk_count; ++chunk_id) {
-          const auto& chunk = table->get_chunk(chunk_id);
-          Assert(chunk->individually_sorted_by().empty(), "Chunk SortColumnDefinitions need to be empty");
-          chunk->set_individually_sorted_by(sort_column);
-        }
+            for (auto chunk_id = ChunkID{0}; chunk_id < chunk_count; ++chunk_id) {
+              const auto& chunk = table->get_chunk(chunk_id);
+              Assert(chunk->individually_sorted_by().empty(), "Chunk SortColumnDefinitions need to be empty");
+              chunk->set_individually_sorted_by(sort_column);
+            }
 
-        continue;
+            return;
+          }
+
+          // We sort the tables after their creation so that we are independent of the order in which they are filled.
+          // For this, we use the sort operator. Because it returns a `const Table`, we need to recreate the table and
+          // migrate the sorted chunks to that table.
+
+          Timer per_table_timer;
+
+          auto table_wrapper = std::make_shared<TableWrapper>(table);
+          table_wrapper->execute();
+          auto sort = std::make_shared<Sort>(
+              table_wrapper, std::vector<SortColumnDefinition>{SortColumnDefinition{sort_column_id, sort_mode}},
+              _benchmark_config->chunk_size, Sort::ForceMaterialization::Yes);
+          sort->execute();
+          const auto immutable_sorted_table = sort->get_output();
+
+          Assert(immutable_sorted_table->chunk_count() == table->chunk_count(), "Mismatching chunk_count");
+
+          table = std::make_shared<Table>(immutable_sorted_table->column_definitions(), TableType::Data,
+                                          table->target_chunk_size(), UseMvcc::Yes);
+          const auto column_count = immutable_sorted_table->column_count();
+          for (auto chunk_id = ChunkID{0}; chunk_id < chunk_count; ++chunk_id) {
+            const auto chunk = immutable_sorted_table->get_chunk(chunk_id);
+            auto mvcc_data = std::make_shared<MvccData>(chunk->size(), CommitID{0});
+            Segments segments{};
+            for (auto column_id = ColumnID{0}; column_id < column_count; ++column_id) {
+              segments.emplace_back(chunk->get_segment(column_id));
+            }
+            table->append_chunk(segments, mvcc_data);
+            table->get_chunk(chunk_id)->finalize();
+            table->get_chunk(chunk_id)->set_individually_sorted_by(SortColumnDefinition(sort_column_id, sort_mode));
+          }
+
+          auto output = std::string{"-  Sorted '"} + table_name + "' by '" + column_name + "' (" +
+                        per_table_timer.lap_formatted() + ")\n";
+          std::cout << output << std::flush;
+        });
       }
+      for (auto& thread : threads) thread.join();
 
-      // We sort the tables after their creation so that we are independent of the order in which they are filled.
-      // For this, we use the sort operator. Because it returns a `const Table`, we need to recreate the table and
-      // migrate the sorted chunks to that table.
-
-      std::cout << "-  Sorting '" << table_name << "' by '" << column_name << "' " << std::flush;
-      Timer per_table_timer;
-
-      auto table_wrapper = std::make_shared<TableWrapper>(table);
-      table_wrapper->execute();
-      auto sort = std::make_shared<Sort>(
-          table_wrapper, std::vector<SortColumnDefinition>{SortColumnDefinition{sort_column_id, sort_mode}},
-          _benchmark_config->chunk_size, Sort::ForceMaterialization::Yes);
-      sort->execute();
-      const auto immutable_sorted_table = sort->get_output();
-
-      Assert(immutable_sorted_table->chunk_count() == table->chunk_count(), "Mismatching chunk_count");
-
-      table = std::make_shared<Table>(immutable_sorted_table->column_definitions(), TableType::Data,
-                                      table->target_chunk_size(), UseMvcc::Yes);
-      const auto column_count = immutable_sorted_table->column_count();
-      for (auto chunk_id = ChunkID{0}; chunk_id < chunk_count; ++chunk_id) {
-        const auto chunk = immutable_sorted_table->get_chunk(chunk_id);
-        auto mvcc_data = std::make_shared<MvccData>(chunk->size(), CommitID{0});
-        Segments segments{};
-        for (auto column_id = ColumnID{0}; column_id < column_count; ++column_id) {
-          segments.emplace_back(chunk->get_segment(column_id));
-        }
-        table->append_chunk(segments, mvcc_data);
-        table->get_chunk(chunk_id)->finalize();
-        table->get_chunk(chunk_id)->set_individually_sorted_by(SortColumnDefinition(sort_column_id, sort_mode));
-      }
-
-      std::cout << "(" << per_table_timer.lap_formatted() << ")" << std::endl;
+      metrics.sort_duration = timer.lap();
+      std::cout << "- Sorting tables done (" << format_duration(metrics.sort_duration) << ")" << std::endl;
     }
-    metrics.sort_duration = timer.lap();
-    std::cout << "- Sorting tables done (" << format_duration(metrics.sort_duration) << ")" << std::endl;
   }
 
   /**
@@ -156,18 +167,28 @@ void AbstractTableGenerator::generate_and_store() {
   /**
    * Encode the tables
    */
-  std::cout << "- Encoding tables (if necessary) and generating pruning statistics" << std::endl;
-  for (auto& [table_name, table_info] : table_info_by_name) {
-    std::cout << "-  Encoding '" << table_name << "' - " << std::flush;
-    Timer per_table_timer;
-    table_info.re_encoded =
-        BenchmarkTableEncoder::encode(table_name, table_info.table, _benchmark_config->encoding_config);
-    std::cout << (table_info.re_encoded ? "encoding applied" : "no encoding necessary");
-    std::cout << " (" << per_table_timer.lap_formatted() << ")" << std::endl;
+  {
+    std::cout << "- Encoding tables (if necessary) and generating pruning statistics" << std::endl;
+
+    auto threads = std::vector<std::thread>{};
+    for (auto& [table_name, table_info] : table_info_by_name) {
+      threads.emplace_back([&] {
+        Timer per_table_timer;
+        table_info.re_encoded =
+            BenchmarkTableEncoder::encode(table_name, table_info.table, _benchmark_config->encoding_config);
+        auto output = std::string{"-  Encoding '"} + table_name + "' - " +
+                      (table_info.re_encoded ? "encoding applied" : "no encoding necessary") + " (" +
+                      per_table_timer.lap_formatted() + ")\n";
+        std::cout << output << std::flush;
+      });
+    }
+
+    for (auto& thread : threads) thread.join();
+
+    metrics.encoding_duration = timer.lap();
+    std::cout << "- Encoding tables and generating pruning statistic done ("
+              << format_duration(metrics.encoding_duration) << ")" << std::endl;
   }
-  metrics.encoding_duration = timer.lap();
-  std::cout << "- Encoding tables and generating pruning statistic done (" << format_duration(metrics.encoding_duration)
-            << ")" << std::endl;
 
   /**
    * Write the Tables into binary files if required
@@ -210,20 +231,27 @@ void AbstractTableGenerator::generate_and_store() {
   /**
    * Add the Tables to the StorageManager
    */
-  std::cout << "- Adding tables to StorageManager and generating table statistics" << std::endl;
-  auto& storage_manager = Hyrise::get().storage_manager;
-  for (auto& [table_name, table_info] : table_info_by_name) {
-    std::cout << "-  Adding '" << table_name << "' " << std::flush;
-    Timer per_table_timer;
-    if (storage_manager.has_table(table_name)) storage_manager.drop_table(table_name);
-    storage_manager.add_table(table_name, table_info.table);
-    std::cout << "(" << per_table_timer.lap_formatted() << ")" << std::endl;
+  {
+    std::cout << "- Adding tables to StorageManager and generating table statistics" << std::endl;
+    auto& storage_manager = Hyrise::get().storage_manager;
+    auto threads = std::vector<std::thread>{};
+    for (auto& [table_name, table_info] : table_info_by_name) {
+      threads.emplace_back([&] {
+        Timer per_table_timer;
+        if (storage_manager.has_table(table_name)) storage_manager.drop_table(table_name);
+        storage_manager.add_table(table_name, table_info.table);
+        const auto output =
+            std::string{"-  Added '"} + table_name + "' " + "(" + per_table_timer.lap_formatted() + ")\n";
+        std::cout << output << std::flush;
+      });
+    }
+    for (auto& thread : threads) thread.join();
+
+    metrics.store_duration = timer.lap();
+
+    std::cout << "- Adding tables to StorageManager and generating table statistics done ("
+              << format_duration(metrics.store_duration) << ")" << std::endl;
   }
-
-  metrics.store_duration = timer.lap();
-
-  std::cout << "- Adding tables to StorageManager and generating table statistics done ("
-            << format_duration(metrics.store_duration) << ")" << std::endl;
 
   /**
    * Create indexes if requested by the user