--- a/src/benchmarklib/abstract_table_generator.cpp
+++ b/src/benchmarklib/abstract_table_generator.cpp
@@ -94,8 +94,9 @@ void AbstractTableGenerator::generate_and_store() {
           });
 
           if (is_sorted) {
-            auto output = std::string{"-  Table '"} + table_name + "' is already sorted by '" + column_name + "'\n";
-            std::cout << output << std::flush;
+            auto output = std::stringstream;
+            output << "-  Table '" << table_name << "' is already sorted by '" << column_name << "'\n";
+            std::cout << output.str() << std::flush;
             const SortColumnDefinition sort_column{sort_column_id, sort_mode};
 
             if (_all_chunks_sorted_by(table, sort_column)) return;