--- a/src/lib/CMakeLists.txt
+++ b/src/lib/CMakeLists.txt
@@ -644,6 +644,7 @@ set(
 set(
     LIBRARIES
     lz4
+    robin_map
     sqlparser
     uninitialized_vector
     zstd