--- a/src/lib/statistics/statistics_objects/equal_distinct_count_histogram.cpp
+++ b/src/lib/statistics/statistics_objects/equal_distinct_count_histogram.cpp
@@ -7,7 +7,7 @@
 #include <utility>
 #include <vector>
 
-#include <robin_map.h>
+#include <robin_map.h>  // NOLINT
 
 #include "generic_histogram.hpp"
 #include "resolve_type.hpp"