--- a/src/lib/statistics/table_statistics.cpp
+++ b/src/lib/statistics/table_statistics.cpp
@@ -24,7 +24,8 @@ std::shared_ptr<TableStatistics> TableStatistics::from_table(const Table& table)
   auto threads = std::vector<std::thread>{};
 
   /**
-   * Parallely create statistics objects for the Table's columns
+   * Parallely create statistics objects for the Table's columns. Do not use JobTask as we want this to be parallel
+   * even if Hyrise is running without a scheduler.
    */
   for (auto thread_id = 0u;
        thread_id < std::min(static_cast<uint>(table.column_count()), std::thread::hardware_concurrency() + 1);