* For the generation of the table histograms, use the robin_map. It allows us to cache the hash of strings, which makes generation much faster. We will likely replace the current bytell hash map with this one for other things, too.
* Parallelize some steps of the benchmark start. Hide whitespaces when reviewing this.

Old:
- Sorting tables done (5 s 957 ms)
- Encoding tables and generating pruning statistic done (10 s 692 ms)
- Adding tables to StorageManager and generating table statistics done (2 min 41 s)
- Total time from binary start to first query: 3 min 21 s

New:
- Sorting tables done (5 s 238 ms)
- Encoding tables and generating pruning statistic done (8 s 778 ms)
- Adding tables to StorageManager and generating table statistics done (1 min 23 s)
- Total time from binary start to first query: 1 min 50 s