@mrks 2020-10-19
Thanks for the quick review.

Clang is right, this is a [defect](http://www.open-std.org/jtc1/sc22/wg21/docs/cwg_defects.html#2313) in C++.